float std of int

+++ 0-add.py
@@ -5,6 +5,6 @@ that at two int and return a float
 """
 
 
-def add(a: int, b: int) -> float:
+def add(a: float, b: float) -> float:
     """This function does the sum of two int"""
     return a + b
